Removed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@7404 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/states_screens/race_gui.cpp
@@ -806,8 +806,8 @@ void RaceGUI::drawSpeedAndEnergy(const Kart* kart, const core::recti &viewport,
     int meter_height       = (int)(SPEEDWIDTH*minRatio);
 
     drawEnergyMeter(viewport.LowerRightCorner.X, 
-                    viewport.LowerRightCorner.Y - meter_height*0.75f, kart,
-                    viewport, scaling);
+                    (int)(viewport.LowerRightCorner.Y - meter_height*0.75f), 
+                    kart, viewport, scaling);
     
     // First draw the meter (i.e. the background which contains the numbers etc.
     // -------------------------------------------------------------------------
